PC-55 Convert owners to JPA entities.
