+++ pet-clinic-data/src/main/java/com/sheffmachine/sfgpetclinic/model/Pet.java
@@ -1,12 +1,21 @@
 package com.sheffmachine.sfgpetclinic.model;
 
+import javax.persistence.*;
 import java.time.LocalDate;
 
+@Entity
+@Table(name = "pets")
 public class Pet extends BaseEntity {
+    @ManyToOne
+    @JoinColumn(name = "type_id")
     private PetType petType;
+    @ManyToOne
+    @JoinColumn(name = "owner_id")
     private Owner owner;
+    @Column(name = "birth_date")
     private LocalDate birthDate;
     private Long id;
+    @Column(name = "name")
     private String name;
 
     public String getName() {
