+++ pet-clinic-data/src/main/java/com/sheffmachine/sfgpetclinic/model/PetType.java
@@ -1,6 +1,13 @@
 package com.sheffmachine.sfgpetclinic.model;
 
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.Table;
+
+@Entity
+@Table(name = "types")
 public class PetType extends BaseEntity {
+    @Column(name = "name")
     private String name;
 
     public String getName() {
